Fix issue calling onSimulationStop of all the UIForces
The `onSimulationStop` call was made on every single UIForce the UIForce controller had, but the actor or force controller not always had all the forces, making calls to undefined values. So I modified the for each loop to take only the forces in the force controller instead of the UI force controller, fixing the issue. Need further testing when all forces are ready with the `onSimulationStop` call for UI and force Management.